Eliminate insane Object allocation rate
The BlockState object had a lot of problems...

Translating a single chunk caused on allocation of at least 65536
objects. Joining a server and walking forwards a bit and taking a jmap
histogram revealed the allocation of ~70 million BlockState objects.
Even worse the replacement was run on boxed Integers of which ~110
million instances were created just because one player moved forward a
few blocks.
When this plugin is used on a reasonably big server the GC pressure is
extremely high.
This patch completely eliminates the use of the BlockState and replaces
the Map with a primitive fastutil map.
After this patch running a few blocks forward only allocates ~12000
boxed Integer objects, which proves that the Chunk replacement was also
the root cause for the huge amount of boxed integers.